requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
fastapi==0.115.8
h11==0.14.0
idna==3.10
limits==4.0.1
lxml==5.3.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
tiktoken==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
